Refactor: modularize API handler into internal packages
Decomposed the monolithic `api/index.go` into domain-specific internal packages to improve maintainability, testability, and separation of concerns.

- `internal/transport`: HTTP client configuration and SSRF protection (`NewSafeClient`).
- `internal/article`: Article fetching and parsing logic (`Fetch`).
- `internal/request`: Request parsing, validation, and URL reconstruction (`NormalizeURL`, `ReconstructURL`, `GetFormat`).
- `internal/formatter`: Output formatting and rendering (`Render`, `Template`).
- Updated `api/index.go` to serve as a clean orchestration layer using these packages.
- Migrated and split tests into their respective package test files.
- Added integration test in `api/index_test.go` to verify wiring.

This refactoring adheres to the Single Responsibility Principle and makes the codebase easier to navigate and extend.
